fix(eslint-errors): modal
