+++ packages/modal/src/confirm/context.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { DialogContextType } from './types';
+
+export const ConfirmationDialogContext = React.createContext<DialogContextType>({} as DialogContextType);
