+++ packages/modal/src/confirm/index.ts
@@ -1,1 +1,5 @@
-export * from './confirm';
+export * from './use-confirm';
+export * from './use-dialog-show';
+export * from './provider';
+export * from './context';
+export type * from './types';
