+++ packages/modal/src/confirm/use-confirm.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { DialogContextType } from './types';
+import { ConfirmationDialogContext } from './context';
+
+const useConfirm = (): DialogContextType => React.useContext(ConfirmationDialogContext);
+
+export { useConfirm };
+export default useConfirm;
